backend/requirements.txt
fastapi
uvicorn
psycopg2-binary>=2.9.9
python-dotenv
yfinance
fear-and-greed
openpyxl
beautifulsoup4
openai
